+++ PetStoreTests/Models/Pet.cs
@@ -10,7 +10,7 @@
         public Category Category { get; set; }
 
         [JsonPropertyName("id")]
-        public int? Id { get; set; }
+        public ulong? Id { get; set; }
 
         [JsonPropertyName("name")]
         public string Name { get; set; }
