+++ PetStoreTests/Models/Tag.cs
@@ -7,7 +7,7 @@
         #region Properties
 
         [JsonPropertyName("id")]
-        public int? Id { get; set; }
+        public ulong? Id { get; set; }
 
         [JsonPropertyName("name")]
         public string Name { get; set; }
